requirements.txt

langchain
langchain-openai
yfinance
pandas
numpy
fastapi
uvicorn
python-dotenv
ta
